requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9
Pillow>=10.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
